pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "emsigma"
version = "0.2.2"
description = "Spectral interpretation using Gaussian mixtures and autoencoder"
readme = "README.md"
requires-python = ">=3.11"
license = "GPL-3.0-or-later"
authors = [
    { name = "Po-Yen Tung", email = "pyt21@cam.ac.uk" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "notebook>=7.5.0",
    "torch>=2.9.0",
    "hyperspy>=2.3.0",
    "exspy>=0.3.0",
    "ipywidgets>=8.1.8",
    "matplotlib>=3.10.0",
    "numpy>=2.0.0",
    "numba>=0.9.0",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.0",
    "umap-learn>=0.5.9",
    "tqdm>=4.67.0",
    "seaborn>=0.13.2",
    "plotly>=6.5.0",
    "altair>=6.0.0",
    "jupyterlab>=4.5.0",
    "pandas>=2.0.0",
    "scikit-image>=0.22.0",
    "pillow>=10.0.0",
    "ipython>=8.0.0",
    "sparse>=0.17.0",
]

[tool.hatch.build.targets.wheel]
packages = ["sigma"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "N", "UP", "PL"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pyright]
include = ["sigma"]
pythonVersion = "3.11"
typeCheckingMode = "basic"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore:.*append2pathname.*:UserWarning",
    "ignore:.*incremental_filename.*:UserWarning",
    "ignore:.*ensure_directory.*:UserWarning",
    "ignore:.*overwrite.*:UserWarning",
    "ignore:.*get_file_handle.*:UserWarning",
    "ignore:.*rsciio.utils.rgb_tools.*:UserWarning",
    "ignore:.*scipy.odr.*:DeprecationWarning",
    "ignore:.*moved to `rsciio.utils.path`.*:Warning",
    "ignore:.*renamed to `rsciio.utils.rgb`.*:Warning",
    "ignore:.*moved to `rsciio.utils.file`.*:Warning",
]

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "ruff>=0.14.11",
]
